Struct solarxr::Data

pub struct Data { /* private fields */ }

The self-referencing struct.

Implementations

impl Data

pub(super) fn new(
data: Vec<u8>,
table_builder: impl for<'this> FnOnce(&'this Vec<u8>) -> MessageBundle<'this>
) -> Data

Constructs a new instance of this self-referential struct. (See also DataBuilder::build()). Each argument is a field of the new struct. Fields that refer to other fields inside the struct are initialized using functions instead of directly passing their value. The arguments are as follows:

Argument	Suggested Use

data	Directly pass in the value this field should contain

table_builder	Use a function or closure: (data: &_) -> table: _

pub(super) async fn new_async(
data: Vec<u8>,
table_builder: impl for<'this> FnOnce(&'this Vec<u8>) -> Pin<Box<dyn Future<Output = MessageBundle<'this>> + 'this>>
) -> Data

Constructs a new instance of this self-referential struct. (See also DataAsyncBuilder::build()). Each argument is a field of the new struct. Fields that refer to other fields inside the struct are initialized using functions instead of directly passing their value. The arguments are as follows:

Argument	Suggested Use

data	Directly pass in the value this field should contain

table_builder	Use a function or closure: (data: &_) -> table: _

pub(super) async fn new_async_send(
data: Vec<u8>,
table_builder: impl for<'this> FnOnce(&'this Vec<u8>) -> Pin<Box<dyn Future<Output = MessageBundle<'this>> + Send + 'this>>
) -> Data

Constructs a new instance of this self-referential struct. (See also DataAsyncSendBuilder::build()). Each argument is a field of the new struct. Fields that refer to other fields inside the struct are initialized using functions instead of directly passing their value. The arguments are as follows:

Argument	Suggested Use

data	Directly pass in the value this field should contain

table_builder	Use a function or closure: (data: &_) -> table: _

pub(super) fn try_new<Error_>(
data: Vec<u8>,
table_builder: impl for<'this> FnOnce(&'this Vec<u8>) -> Result<MessageBundle<'this>, Error_>
) -> Result<Data, Error_>

(See also DataTryBuilder::try_build().) Like new, but builders for self-referencing fields can return results. If any of them fail, Err is returned. If all of them succeed, Ok is returned. The arguments are as follows:

Argument	Suggested Use

data	Directly pass in the value this field should contain

table_builder	Use a function or closure: (data: &_) -> Result<table: _, Error_>

pub(super) fn try_new_or_recover<Error_>(
data: Vec<u8>,
table_builder: impl for<'this> FnOnce(&'this Vec<u8>) -> Result<MessageBundle<'this>, Error_>
) -> Result<Data, (Error_, Heads)>

(See also DataTryBuilder::try_build_or_recover().) Like try_new, but all head fields are returned in the case of an error. The arguments are as follows:

Argument	Suggested Use

data	Directly pass in the value this field should contain

table_builder	Use a function or closure: (data: &_) -> Result<table: _, Error_>

pub(super) async fn try_new_async<Error_>(
data: Vec<u8>,
table_builder: impl for<'this> FnOnce(&'this Vec<u8>) -> Pin<Box<dyn Future<Output = Result<MessageBundle<'this>, Error_>> + 'this>>
) -> Result<Data, Error_>

(See also DataAsyncTryBuilder::try_build().) Like new, but builders for self-referencing fields can return results. If any of them fail, Err is returned. If all of them succeed, Ok is returned. The arguments are as follows:

Argument	Suggested Use

data	Directly pass in the value this field should contain

table_builder	Use a function or closure: (data: &_) -> Result<table: _, Error_>

pub(super) async fn try_new_or_recover_async<Error_>(
data: Vec<u8>,
table_builder: impl for<'this> FnOnce(&'this Vec<u8>) -> Pin<Box<dyn Future<Output = Result<MessageBundle<'this>, Error_>> + 'this>>
) -> Result<Data, (Error_, Heads)>

(See also DataAsyncTryBuilder::try_build_or_recover().) Like try_new, but all head fields are returned in the case of an error. The arguments are as follows:

Argument	Suggested Use

data	Directly pass in the value this field should contain

table_builder	Use a function or closure: (data: &_) -> Result<table: _, Error_>

pub(super) async fn try_new_async_send<Error_>(
data: Vec<u8>,
table_builder: impl for<'this> FnOnce(&'this Vec<u8>) -> Pin<Box<dyn Future<Output = Result<MessageBundle<'this>, Error_>> + Send + 'this>>
) -> Result<Data, Error_>

(See also DataAsyncSendTryBuilder::try_build().) Like new, but builders for self-referencing fields can return results. If any of them fail, Err is returned. If all of them succeed, Ok is returned. The arguments are as follows:

Argument	Suggested Use

data	Directly pass in the value this field should contain

table_builder	Use a function or closure: (data: &_) -> Result<table: _, Error_>

pub(super) async fn try_new_or_recover_async_send<Error_>(
data: Vec<u8>,
table_builder: impl for<'this> FnOnce(&'this Vec<u8>) -> Pin<Box<dyn Future<Output = Result<MessageBundle<'this>, Error_>> + Send + 'this>>
) -> Result<Data, (Error_, Heads)>

(See also DataAsyncSendTryBuilder::try_build_or_recover().) Like try_new, but all head fields are returned in the case of an error. The arguments are as follows:

Argument	Suggested Use

data	Directly pass in the value this field should contain

table_builder	Use a function or closure: (data: &_) -> Result<table: _, Error_>

pub(super) fn with_data<'outer_borrow, ReturnType>(
&'outer_borrow self,
user: impl for<'this> FnOnce(&'outer_borrow Vec<u8>) -> ReturnType
) -> ReturnType

Provides limited immutable access to data. This method was generated because the contents of data are immutably borrowed by other fields.

pub(super) fn borrow_data<'this>(&'this self) -> &'this Vec<u8>

Provides limited immutable access to data. This method was generated because the contents of data are immutably borrowed by other fields.

pub(super) fn with_table<'outer_borrow, ReturnType>(
&'outer_borrow self,
user: impl for<'this> FnOnce(&'outer_borrow MessageBundle<'this>) -> ReturnType
) -> ReturnType

Provides an immutable reference to table. This method was generated because table is a tail field.

pub(super) fn borrow_table<'this>(&'this self) -> &'this MessageBundle<'this>

Provides an immutable reference to table. This method was generated because table is a tail field.

pub(super) fn with_table_mut<'outer_borrow, ReturnType>(
&'outer_borrow mut self,
user: impl for<'this> FnOnce(&'outer_borrow mut MessageBundle<'this>) -> ReturnType
) -> ReturnType

Provides a mutable reference to table. This method was generated because table is a tail field. No borrow_table_mut function was generated because Rust’s borrow checker is currently unable to guarantee that such a method would be used safely.

pub(super) fn with<'outer_borrow, ReturnType>(
&'outer_borrow self,
user: impl for<'this> FnOnce(BorrowedFields<'outer_borrow, 'this>) -> ReturnType
) -> ReturnType

This method provides immutable references to all tail and immutably borrowed fields.

pub(super) fn with_mut<'outer_borrow, ReturnType>(
&'outer_borrow mut self,
user: impl for<'this> FnOnce(BorrowedMutFields<'outer_borrow, 'this>) -> ReturnType
) -> ReturnType

This method provides mutable references to all tail fields.

pub(super) fn into_heads(self) -> Heads

This function drops all internally referencing fields and returns only the head fields of this struct.

impl Data

pub unsafe fn from_vec_unchecked(data: Vec<u8>) -> Self

Safety

data must have already been verified to be a valid [MessageBundle] flatbuffer

pub fn from_vec(data: Vec<u8>) -> Result<Self, (Vec<u8>, DecodeError)>

pub fn into_vec(self) -> Vec<u8>

pub fn as_slice(&self) -> &[u8]

pub fn table(&self) -> MessageBundle<'_>